Refactored notification configuration endpoints into its own controller
